add a filter_and_sort_categories function
